challenges/web/This_WEB_Too_Simple/service/requirements.txt
Flask
PyJWT
pytz
